Extract AttrMasker models' class methods
Move them to a separate module.  Keep the AttrMasker module clean and
minimal.
